+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - 0.6.2
     - Filtered and sorted rule condition paths to be more user friendly
     - Added "chamber" fact (currentTime, remainingTime, anyWizardDead, enemiesRemaining)
+    - Added "target wizard" and "target enemy" to actions
 - 0.6.1
     - Added magizoologist skill tree
 - 0.6.0